Client/release/1.0.0 (#5)
* client: fix: reubicacion de layout

* client: feat: navegacion para usuarios conectados

* client: deploy en vercel
