+++ client/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import { Header } from "@/components/layout/header";
-import { Footer } from "@/components/layout/footer";
+import { Header } from "@/layouts/header";
+import { Footer } from "@/layouts/footer";
 import "./globals.css";
 import { Suspense } from "react";
